move pricing pages to their own file
