+++ app/main/__init__.py
@@ -29,6 +29,7 @@ from app.main.views import (  # noqa isort:skip
     organisations,
     performance,
     platform_admin,
+    pricing,
     providers,
     register,
     returned_letters,
@@ -37,8 +38,8 @@ from app.main.views import (  # noqa isort:skip
     sign_in,
     sign_out,
     templates,
-    two_factor,
     tour,
+    two_factor,
     uploads,
     user_profile,
     verify,
