+++ app/main/views/index.py
@@ -1,6 +1,5 @@
 from flask import (
     abort,
-    current_app,
     make_response,
     redirect,
     render_template,
@@ -8,20 +7,15 @@ from flask import (
     url_for,
 )
 from flask_login import current_user
-from notifications_utils.international_billing_rates import (
-    INTERNATIONAL_BILLING_RATES,
-)
 from notifications_utils.template import HTMLEmailTemplate, LetterImageTemplate
 
 from app import email_branding_client, letter_branding_client, status_api_client
 from app.main import main
-from app.main.forms import FieldWithNoneOption, SearchByNameForm
+from app.main.forms import FieldWithNoneOption
 from app.main.views.sub_navigation_dictionaries import (
     features_nav,
-    pricing_nav,
     using_notify_nav,
 )
-from app.models.feedback import QUESTION_TICKET_TYPE
 from app.utils import get_logo_cdn_domain, hide_from_search_engines
 
 
@@ -57,43 +51,6 @@ def privacy():
 @main.route('/accessibility-statement')
 def accessibility_statement():
     return render_template('views/accessibility_statement.html')
-
-
-@main.route('/pricing')
-def pricing():
-    return render_template(
-        'views/pricing/index.html',
-        sms_rate=0.0158,
-        international_sms_rates=sorted([
-            (cc, country['names'], country['billable_units'])
-            for cc, country in INTERNATIONAL_BILLING_RATES.items()
-        ], key=lambda x: x[0]),
-        search_form=SearchByNameForm(),
-        navigation_links=pricing_nav(),
-    )
-
-
-@main.route('/pricing/how-to-pay')
-def how_to_pay():
-    return render_template(
-        'views/pricing/how-to-pay.html',
-        support_link=url_for('main.feedback', ticket_type=QUESTION_TICKET_TYPE),
-        navigation_links=pricing_nav(),
-    )
-
-
-@main.route('/pricing/billing-details')
-def billing_details():
-    if current_user.is_authenticated:
-        return render_template(
-            'views/pricing/billing-details.html',
-            billing_details=current_app.config['NOTIFY_BILLING_DETAILS'],
-            navigation_links=pricing_nav(),
-        )
-    return render_template(
-        'views/pricing/billing-details-signed-out.html',
-        navigation_links=pricing_nav(),
-    )
 
 
 @main.route('/delivery-and-failure')
