+++ app/templates/views/pricing/billing-details.html
@@ -16,7 +16,7 @@
     </p>
 
     <p class="govuk-body">
-      <a class="govuk-link govuk-link--no-visited-state" href="{{ support_link }}">Contact us</a> if you need any other details.
+      <a class="govuk-link govuk-link--no-visited-state" href="{{ url_for('main.support') }}">Contact us</a> if you need any other details.
     </p>
 
     <h2 class="heading-medium govuk-!-margin-top-7">
