+++ app/templates/views/pricing/how-to-pay.html
@@ -53,13 +53,13 @@
     <p class="govuk-body">
       Your organisation should raise a single <abbr title="purchase order">PO</abbr> for the estimated cost of all its services. You can update the <abbr title="purchase order">PO</abbr> any time if your usage increases.
     </p>
- 
+
     <p class="govuk-body">
       Your organisation may need to <a class="govuk-link govuk-link--no-visited-state" href="{{ url_for('main.billing_details') }}">add the Cabinet Office as a supplier</a> before you can raise a <abbr title="purchase order">PO</abbr>.
     </p>
 
     <p class="govuk-body">
-       <a class="govuk-link govuk-link--no-visited-state" href="{{ support_link }}">Contact us</a> if you need help raising a purchase order.
+       <a class="govuk-link govuk-link--no-visited-state" href="{{ url_for('main.support') }}">Contact us</a> if you need help raising a purchase order.
     </p>
 
 {% endblock %}
